Split MCP server and TUI into per-feature modules
Extract monolithic mcp-server.ts (786→60 lines) and tui.ts (1442→120 lines) into per-feature modules. Each data source (Notes, Messages, Contacts) now owns its MCP tool registrations and TUI code in its own directory. Shared helpers are DRYed up: highlightLine, scrollIntoView, handleScrollKeys, buildFooter, and initDb replace repeated patterns across all three features.
